:sparkles: refactor project component
to render individual project items, instead of the whole list.  Pushes the listing to the parent component, which leaves it flexible enough to also be used in the home page carousel.  Also adds the carousel to the home page.
