+++ src/components/projectList.scss
@@ -1,6 +1,12 @@
 .projects {
     position: relative;
 
+    &-items {
+        display: grid;
+        grid-template-columns: 1fr 1fr 1fr 1fr 1fr 1fr;
+        width: 960px;
+    }
+
     &-search {
         position: absolute;
         right: 140px;
