+++ src/pages/index.js
@@ -3,10 +3,13 @@ import { graphql, Link } from 'gatsby';
 import { Helmet } from 'react-helmet';
 import PropTypes from 'prop-types';
 
+import { CarouselProvider, Slide, Slider } from 'pure-react-carousel';
+import 'pure-react-carousel/dist/react-carousel.es.css';
+
 import './index.scss';
 
 import Layout from '../components/layout';
-import ProjectBoxes from '../components/shared/projectBoxes';
+import ProjectBox from '../components/shared/projectBox';
 import Statement from '../components/vision/statement';
 
 const Index = ({
@@ -23,7 +26,25 @@ const Index = ({
       <Statement home />
       <div>
         {projectsNodes.length > 0 && <h3>Featured Projects</h3>}
-        <ProjectBoxes projects={projectsNodes} />
+        <CarouselProvider
+          totalSlides={projectsNodes.length}
+          visibleSlides={6}
+          step={6}
+          naturalSlideWidth={140}
+          naturalSlideHeight={182}
+          infinite
+          isPlaying
+          interval={5000}
+        >
+          <Slider>
+            {projectsNodes.map((project) => (
+              <Slide key={project.id}>
+                <ProjectBox node={project} />
+              </Slide>
+            ))}
+          </Slider>
+        </CarouselProvider>
+
         <h3 className="more-projects"><Link to="/projects/">See more projects +</Link></h3>
       </div>
     </Layout>
