+++ src/pages/person/{PeopleJson.fieldData__slug}.js
@@ -5,7 +5,7 @@ import PropTypes from 'prop-types';
 
 import Layout from '../../components/layout';
 import NormalizeP from '../../components/filters/normalizeP';
-import ProjectBoxes from '../../components/shared/projectBoxes';
+import ProjectBox from '../../components/shared/projectBox';
 
 import '../detailPage.scss';
 
@@ -22,7 +22,7 @@ const Person = ({ data: { allPeopleJson: { edges: [{ node }, ...rest] } } }) => 
   // Filter out projects with no slug, they are artifacts from unpublished projects.
   const projects = node.portalData.PeopleProjectJoin_People_WebView.filter(({ Projects_People_WebView__slug: slug }) => slug !== '');
 
-  // Format portal data to match the format of the shared ProjectBoxes component.
+  // Format portal data to match the format of the shared ProjectBox component.
   const projectsNodes = projects.map((project) => (
     {
       id: project.Projects_People_WebView__slug,
@@ -51,8 +51,12 @@ const Person = ({ data: { allPeopleJson: { edges: [{ node }, ...rest] } } }) => 
         <img className="detail-badge" alt={FullName} src={`${process.env.GATSBY_MEDIA_LIBRARY}/${cBadgeRawURL}?width=140`} />
         <NormalizeP className="description" mixedMarkup={DescriptionHTML} />
       </div>
+
       {projectsNodes.length > 0 && <h3>Projects</h3>}
-      <ProjectBoxes projects={projectsNodes} />
+      <div className="projects-items">
+        {projectsNodes.map((project) => <ProjectBox key={project.id} node={project} />)}
+      </div>
+
       {papers.length > 0 && <h3 className="papers-title">Papers</h3>}
       {papers.map((paper) => (
         <div key={paper.recordId}>
